requirements.txt
langchain>=0.1.0
langchain-core>=0.1.15
langgraph>=0.0.19
neo4j>=5.13.0
openai>=1.6.0
python-dotenv>=1.0.0
sentence-transformers>=2.2.2
gradio>=4.11.0
huggingface_hub>=0.19.0
unstructured>=0.11.0
chromadb>=0.4.18 